requirements.txt
asgiref==3.8.1
defusedxml==0.7.1
diff-match-patch==20241021
Django==5.2.3
django-import-export==4.3.8
django-unfold==0.62.0
django-vite==3.1.0
et_xmlfile==2.0.0
gunicorn==23.0.0
numpy==2.3.1
odfpy==1.4.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
tabulate==0.9.0
tzdata==2025.2
xlrd==2.0.2
xlwt==1.3.0
